ai-bridge/requirements.txt
flask
openai
feedparser
redis
requests
python-dotenv
gunicorn

# google-genai